# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Threat Modeling Platform API
  description: |
    # World-Class Threat Modeling Platform API
    
    This is a comprehensive API for an enterprise-grade threat modeling platform that provides:
    
    ## 🚀 Key Features
    - **TMAC (Threat Modeling as Code)** - Define threat models as code
    - **AI-Powered Analysis** - Automated threat identification and risk assessment
    - **Multiple Methodologies** - STRIDE, PASTA, LINDDUN, VAST, and more
    - **Enterprise Integration** - SSO, SIEM, DevOps pipeline integration
    - **Real-time Collaboration** - Multi-user editing and review workflows
    - **Comprehensive Reporting** - Executive summaries to technical details
    
    ## 🔒 Security
    - JWT-based authentication with refresh tokens
    - Role-based access control (RBAC)
    - Rate limiting and DDoS protection
    - CORS security with allowlisted origins
    - Input validation and sanitization
    
    ## 📊 Monitoring
    - Health checks for all services
    - Prometheus metrics
    - Distributed tracing
    - Comprehensive logging
    
    ## 🔧 Service Architecture
    - **API Gateway** (Port 3000) - Central routing and security
    - **Auth Service** (Port 3001) - Authentication and authorization
    - **Core Service** (Port 3002) - Threat models and projects
    - **AI Service** (Port 3003) - ML-powered threat analysis
    - **Diagram Service** (Port 3004) - Visual threat modeling
...
    
  version: '1.0.0'
  contact:
...
  license:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/sso/templates/generate:
    post:
      tags: [SSO Management]
      summary: Generate SSO provider configuration
      description: |
        Generate a pre-configured SSO provider template based on the provider type.
        
        **Supported Providers:**
        - azure-ad (requires: tenantId, clientId, clientSecret)
        - google-workspace (requires: clientId, clientSecret, hostedDomain)
        - okta (requires: oktaDomain, clientId, clientSecret)
        - saml (requires: ssoUrl, x509Certificate)
        - oidc (requires: issuer, clientId, clientSecret)
        - ping-identity (requires: environmentId, clientId, clientSecret)
      operationId: generateSSOTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SSOTemplateRequest'
            examples:
              azure_ad:
                summary: Azure AD template
                value:
                  provider: "azure-ad"
                  domain: "threatmodeling.company.com"
                  tenantId: "12345678-1234-1234-1234-123456789012"
                  clientId: "87654321-4321-4321-4321-210987654321"
                  clientSecret: "your-client-secret"
              okta:
                summary: Okta template
                value:
                  provider: "okta"
                  domain: "threatmodeling.company.com"
                  oktaDomain: "company.okta.com"
                  clientId: "okta-client-id"
                  clientSecret: "okta-client-secret"
      responses:
        '200':
          description: SSO template generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSOTemplateResponse'
  /auth/sso/wizard/{provider}:
    get:
      tags: [SSO Management]
      summary: Get configuration wizard steps
      description: |
        Get step-by-step configuration wizard for the specified SSO provider.
        
        **Wizard Features:**
        - Step-by-step configuration guide
        - Field validation and descriptions
        - Security recommendations
        - Estimated completion time
      operationId: getSSOWizard
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [azure-ad, google-workspace, okta, saml, oidc, ping-identity]
          description: SSO provider type
      responses:
        '200':
          description: Configuration wizard steps retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSOWizardResponse'
  /auth/sso/analytics:
    get:
      tags: [SSO Management]
      summary: Get SSO analytics and metrics
      description: |
        Comprehensive SSO analytics including performance metrics, security reports, and trends.
        
        **Analytics Include:**
        - Login performance metrics (avg, P95, P99)
        - Security statistics and threat detection
        - Provider-specific metrics
        - Trend analysis and growth metrics
      operationId: getSSOAnalytics
      parameters:
        - name: timeRange
          in: query
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
          description: Time range for analytics
      responses:
        '200':
          description: SSO analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSOAnalyticsResponse'
  /auth/sso/alerts:
    get:
      tags: [SSO Management]
      summary: Get SSO security alerts
      description: |
        Get SSO security and operational alerts with filtering capabilities.
        
        **Alert Types:**
        - Security: Failed logins, suspicious activity
        - Performance: Slow response times, timeouts
        - Availability: Provider downtime, certificate issues
        - Compliance: Policy violations, audit requirements
      operationId: getSSOAlerts
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [security, performance, availability, compliance]
          description: Filter by alert type
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
          description: Filter by alert severity
        - name: acknowledged
          in: query
          schema:
            type: boolean
          description: Filter by acknowledgment status
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of alerts to return
      responses:
        '200':
          description: SSO alerts retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSOAlertsResponse'
  /auth/sso/alerts/{alertId}/acknowledge:
    post:
      tags: [SSO Management]
      summary: Acknowledge SSO alert
      description: Mark an SSO alert as acknowledged to prevent continued notifications
      operationId: acknowledgeSSOAlert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
          description: Alert ID to acknowledge
      responses:
        '200':
          description: Alert acknowledged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/sso/test-configuration:
    post:
      tags: [SSO Management]
      summary: Test SSO provider configuration
      description: |
        Test and validate SSO provider configuration before activation.
        
        **Testing Includes:**
        - Configuration validation
        - Connectivity testing
        - Certificate verification
        - Attribute mapping validation
        - Security recommendations
      operationId: testSSOConfiguration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SSOProviderConfig'
      responses:
        '200':
          description: Configuration test completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSOTestResponse'
...
  /auth/sso/compliance/report:
    post:
      tags: [SSO Management]
      summary: Generate compliance report
      description: |
        Generate comprehensive compliance report for various standards.
        
        **Supported Standards:**
        - SOC2 Type II
        - GDPR compliance
        - HIPAA requirements
        - PCI-DSS standards
      operationId: generateComplianceReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceReportRequest'
      responses:
        '200':
          description: Compliance report generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceReportResponse'

  # ========================================
  # SYSTEM ENDPOINTS
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        timestamp:
          type: string
          format: date-time
...